Reorganize Rust bindings
