+++ py-polars/src/functions/mod.rs
@@ -1,3 +1,4 @@
+pub mod as_datatype;
 pub mod eager;
 pub mod io;
 pub mod lazy;
